pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "paf-noise-cli"
version = "0.1.0"
description = "CLI per esportare la banca dati rumore del Portale Agenti Fisici"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "python-docx>=1.1.0",
    "mammoth>=1.6.0",
    "PyJWT>=2.12.0",
    "cryptography>=44.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "nh3>=0.3.4",
    "structlog>=25.5.0",
    "fastapi-limiter>=0.2.0",
    "pymupdf>=1.27.0",
    "chromadb>=1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
paf-noise = "paf_noise_cli.cli:main"
create-admin = "src.cli.create_admin:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-k \"not slow\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501", "E712"]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
